little refactoring, save temp profiles

+++ octo/consts.go
@@ -1,4 +1,4 @@
 package octo
 
-const MAX_BED_TEMPERATURE int = 110
-const MAX_TOOL_TEMPERATURE int = 285
+const MAX_BED_TEMPERATURE float64 = 110.0
+const MAX_TOOL_TEMPERATURE float64 = 285.0
